Updated menu and added new scene to make pbj
i just added some QOA updates to the menu. then i made it so when you
get all the ingredients you can go to the next room, in the next room
there a table and the knife now follows your curser. When you  hover
over the loaf it makes 2 slices of bread